Redefined objectives
As I have redefined my objectives and have also spoken to my supervisor about it, I am using a different approach to my individual project. That is I will firstly conduct extensive research on machine learning and then develop a streamlit web app. Firstly, I loaded in the data into the notebook and cleaned the data and created the target which is a classifier to determine if the price goes up or down.